test(bd-7v8.4): extend the node-builtin guard to the webpack bake pair
The guard core moves to a shared helper; the transport-webpack suite
asserts both halves of the bake (producer render.js + consumer.js) carry
no statically-evaluated node builtin import — the consumer bundle exists
precisely so the request path composes on a node:-free runtime.
